[svc-mat-3]: example evaluating math expressions
